[MS-9030] add mock endpoint
